@@ -2,27 +2,26 @@
 
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-
 import 'package:recordify/recordify.dart';
 import 'package:recordify/src/serializer.dart';
 import 'package:recordify/src/recordify_base.dart';
 import 'package:dev_test/test.dart';
-import 'package:reflectable/reflectable.dart';
-import 'package:tekartik_common_utils/common_utils_import.dart';
+
+import 'serializer_test.reflectable.dart';
 
 @Record()
 class Simple {
   @Id()
-  String id;
+  String? id;
 
   //@Ignore()
   //String shouldIgnore = "some_value";
 
-  String stringValue;
-  int intValue;
+  String? stringValue;
+  int? intValue;
 
   @Index()
-  int index;
+  int? index;
 
   void test() {
     print('test');
@@ -30,24 +29,25 @@ class Simple {
 }
 
 main() {
+  initializeReflectable();
   setUpAll(() {});
 
   tearDownAll(() {});
 
   group("serializer", () {
-    solo_test("invokeGetter", () {
-      Simple simple = new Simple()..id ="test";
+    test("invokeGetter", () {
+      Simple simple = Simple()..id = "test";
       //expect(recordify.reflect(simple).invokeGetter("id"), "test");
       //recordify.reflect(simple).invoke('test', []);
       //ClassMirror typeMirror = recordify.reflectType(Simple);
       //simple = typeMirror.newInstance('', []);
-      devPrint(recordify.reflect(simple));
+      var instanceMirror = recordify.reflect(simple);
+      expect(instanceMirror.invokeGetter('id'), 'test');
     });
 
     test("toMap", () {
-      Simple simple = new Simple();
+      Simple simple = Simple();
       expect(toMap(simple), {});
-
 
       simple.id = "test";
       simple.stringValue = "some_string";
@@ -57,14 +57,15 @@ main() {
     });
 
     test("fromMap", () {
-      Simple simple = fromMap({}, Simple);
+      Simple simple = fromMap<Simple>({});
       expect(simple.id, isNull);
 
-      simple = fromMap( {'id': 'test', 'stringValue': 'some_string', 'intValue': 1234}, Simple);
+      simple = fromMap<Simple>(
+        {'id': 'test', 'stringValue': 'some_string', 'intValue': 1234},
+      );
       expect(simple.id, "test");
       expect(simple.stringValue, "some_string");
       expect(simple.intValue, 1234);
-
     });
   });
 }
