@@ -1,14 +1,10 @@
 // Copyright (c) 2016, Alexandre Roux Tekartik. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-import 'package:recordify/recordify.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('A group of tests', () {
-
-    test('First Test', () {
-
-    });
+    test('First Test', () {});
   });
 }
